batch2/restapi/backend/requirements.txt
Django>=3.2
djangorestframework
mysqlclient
gunicorn
